Fix int overflow in DeviceTensor capacity for large-scale PA
DeviceTensor::capacity was declared as `int`, but the constructor
computes the total number of elements as `long int`. The assignment
`capacity = nb` silently truncates when the product of dimensions
exceeds INT_MAX (2^31), causing capacity to become negative. The
subsequent check `(capacity > 0) ? data_ : NULL` then sets data
to NULL, leading to a segfault on first access.

This affects partial assembly (PA) on meshes with >8.8M hexahedral
P1 elements, where the Jacobian Reshape(j_, 3,3,3,3,3,NE) produces
243*NE > INT_MAX.

Reproducer: ParBilinearForm with AssemblyLevel::PARTIAL and
DiffusionIntegrator on a 471x471x40 hex mesh (8,873,640 elements).

Fix: change `int capacity` to `long int capacity` in dtensor.hpp.

+++ linalg/dtensor.hpp
@@ -83,7 +83,7 @@ template<int Dim, typename Scalar = real_t>
 class DeviceTensor
 {
 protected:
-   int capacity;
+   long int capacity;
    Scalar *data;
    int sizes[Dim];
